TCA-55 the ar.scacchipa.twittercloneapp.data.model was created and the files were moved. The login_nav_graph.xml was fixed
