+++ app/src/main/java/ar/scacchipa/twittercloneapp/data/datasource/IAuthDataSource.kt
@@ -1,5 +1,6 @@
 package ar.scacchipa.twittercloneapp.data.datasource
 
+import ar.scacchipa.twittercloneapp.data.model.UserAccessTokenData
 import retrofit2.Response
 import retrofit2.http.Field
 import retrofit2.http.FormUrlEncoded
