+++ app/src/main/java/ar/scacchipa/twittercloneapp/data/model/UserAccessTokenData.kt
@@ -1,4 +1,4 @@
-package ar.scacchipa.twittercloneapp.data.datasource
+package ar.scacchipa.twittercloneapp.data.model
 
 import com.google.gson.annotations.SerializedName
