+++ app/src/test/java/ar/scacchipa/twittercloneapp/data/model/.gitkeep
@@ -1,0 +1,1 @@
+//To track the empty directory
